requirements.txt
asgiref==3.7.2
Django==3.2.20
djangorestframework==3.14.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
djoser==2.2.0
boto3==1.26.142
drf-yasg==1.21.6